Reorganise Marching Cubes and make use of the new device-cuda runtime
The example now makes use of the new device-cuda runtime. The project file
has also been split into seperate device and host projects because the host
code depends on the opengl project which is incompatible with the
device-cuda runtime, and I don't believe there's an easy way to use scenario
variables to work around that. Beside, the device code is conceptionally
different code anyway; ideally we could find a way to to make the device
project a subproject of the host project.

Source code has also been moved around to give the example a bit more
structure.

TN: U114-058