Generate CUDA bindings automatically
With this change the CUDA bindings version is checked at startup and
the corresponding bindings are generated if it does not correspond to
the local informations.

At the moment the only version information is that of gnatvsn, which
is pretty limited.

$CUDA_ROOT should be specified so that the binding script works properly.
By default it uses /usr/local/cuda which is unfit for Windows.