Fix env.sh sourcing
As we are supposed to source the env.sh, $0 is not set to the script name but
rather to the shell path.
This means we cannot use it to find the current script directory, so this
commit reintroduces `pwd` to find the source directory, this adds the
pre-condition that the call must be done from env.sh's own directory.

These hypotheses are checked:

  1. the file is sourced
  2. from its own directory

For the 2nd check, we simply verify that $(pwd)/env.sh exists, it is far from
foolproof but, unless we want to rely on bash-isms this is the only option I
could find.

We also add the -e error shell option