Handle new definition of cudaGetDeviceProperties
Starting with CUDA 12.0, cuda_runtime_api.h defines
cudaGetDeviceProperties as a macro that expands to
cudaGetDeviceProperties_v2. That macro is not handled by
`gcc -fdump-ada-spec`.

This patch adds a UWrap rule that makes it so that the thick
bindings expose a Get_Device_Properties function that calls the
"_v2" version internally, just like the upstream C header does.