Rename sample directory
Starting with version 11.6, CUDA does not come with samples in the
distribution anymore. The recommended source for samples is now
https://github.com/nvidia/cuda-samples.

In the newly recommended repository, the directory hierarchy is
different from the one that came with CUDA. The renaming that this
patch performs makes GNAT for CUDA samples mirror the CUDA samples
from https://github.com/nvidia/cuda-samples.