Fix Makefile dependencies
Device-side build artifacts need to be generated before compilation
of the host side starts. Before the patch, the Makefile rules failed
to meet that requirement, and things only worked when running make
with a single job, i.e. with -j1.

This patch tweaks the Makefile rules so that make is aware of the
fact that the host side should only be compiled after the device
side.