Restrict the scope of "_v2" symbol transformation
A recent change added a UWrap rule to handle preprocessor macros
introduced in CUDA 12.0. Unfortunately, that change broke the
samples with CUDA 11.0.

This patch modifies the UWrap rule to have a much tighter scope.
Instead of trying to handle the general pattern of function
versioning used in upstream CUDA headers, it's now specifically
about the specific function that causes problem with our samples
and CUDA 12.0, cudaGetDeviceProperties.