Disable device-side thick bindings generation
The device-side thick bindings are not usable at the moment, and they
tend to break with each new version of CUDA that NVIDIA releases.
This patch unhooks the generation of these bindings from `setup.sh`.
The uwrap files are left in the tree in the hope that they will be
fixed later.