Remove workaround that's no longer needed
Before this patch, the configuration file for the Sphinx
documentation had a workaround for an issue that's no longer present
with the version of Sphinx that comes with anod. This patch removes
that workaround.

The fact that the issue has been fixed upstream can be checked by
browsing the following GitHub issue:

https://github.com/readthedocs/readthedocs.org/issues/2569