Document CUDA installation restriction
We don't support CUDA installations in /usr because the way we set
PATH then makes the system compiler in `/usr/bin/gcc`, which often
does not support Ada, shadow the compiler driver from GNAT. This
is also true for other directories that contain a installation of gcc.
This patch documents this restriction.